+++ erica/erica_legacy/api/v1/endpoints/grundsteuer.py
@@ -3,7 +3,7 @@ import logging
 from fastapi import status, APIRouter, HTTPException
 
 from erica.erica_legacy.pyeric.eric_errors import EricProcessNotSuccessful
-from erica.application.grundsteuer.grundsteuer_input.grundsteuer_input import GrundsteuerPayload
+from erica.application.grundsteuer.grundsteuer_input import GrundsteuerPayload
 from erica.application.EricRequestProcessing.grundsteuer_request_controller import GrundsteuerRequestController
 
 router = APIRouter()
